feat: Add virtual shard id to fids in the merkle trie to make re-sharding easier (#313)
Add a "virtual" shard id as the first byte of the trie key for messages
so when we need to reshuffle fids among the blockchain shards, we don't
have to move every single fid root individually. We can just remap the
256 virtual shards to actual physical shards.

There was an unexpected complication with an implicit assumption that
the `TIMESTAMP_LENGTH` (really the `UNCOMPACTED_LENGTH`) of the trie
matches the number of nibbles in the trie key for a message type due to
the branching factor.