Support max age of snapshots as a duration
Instead of relying on the block height for expiring snapshots, we'll
instead rely on the age of block timestamps to expire snapshots. This
allows us to avoid issues that may arise when new nodes come online and
sync the chain as we'd open (and subsequently close) a lot of snapshots if
we purely utilized block height as our metric.