feat: Allow multi message simulation with dependent messages (#602)
Allow dependent messages during bulk submit. 

Dependent messages are when one message in the array of messages being
submitted depends on an earlier message in the same array. For eg.,
submitting a `[ENS proof , UserDataAdd]` message[] , which allows both
messages to be accepted and submitted in one go. See the
`test_simulate_bulk_messages_username_proof_and_set` for an example

The main change is here is to pass the DB txn all the way down the
stores, so that they can use the current transaction + DB (instead of
just the DB) to validate/process messages.

This PR allows ENS proofs to be submitted along with their dependent
messages. (Other types of dependent messages (onchain events,
verifications, fnames etc....) are in a separate PR