fix: Handle replicator cursor gets stuck (#617)
There was an issue with the replicator cursor getting stuck, and
returning the same messages over and over again. Somewhat nuanced, but
here's Claude's description of the problem:

The main issue was that the FIDs can have 1 register, but multiple
transfer events. They are all indexed as
OnChainEventType::EventTypeIdRegister on chain events, but this captures
both register and transfer events. In the FidIterator, we're iterating
over this event to get FIDs, but this also inadvertently captures
transfer events, making multiple duplicate FIDs appear in the iterator.
We need to filter more strictly in FidIterator::fetch()