fix: Fix fname secondary DB index (#622)
The secondary index for fnames (To lookup an FID's UserNameProofs) is
missing for about ~1% of FIDs. It's not clear why this is missing, but
it's more common for older FIDs than newer ones, so probably this bug
existed in the code a while ago and is now fixed.

This PR adds a DB migration framework that can do both blocking (ie.
blocks engine startup) and background migrations. It adds the first
migration to fix the secondary index keys for UserNameProofs as a
background migration. It updates the DB version to schema version 2.

Read all the Fnames from the merkle trie, and check that they are
present in the DB and have the correct secondary index.