feat: decouple shard 0 block production from other shards block production (#619)
The approach is
- If the shard witness for the next height after the last recorded shard
witness is available, use it.
- If not, carry over the previous shard witness. 

This makes it easy to figure out what the last recorded shard witness
for each shard is (you can look at the previous block) and doesn't
introduce any complexity around validating multiple shard chunks for a
single block. It also maintains the invariant that the number of shard
witnesses in each block matches the number of shards.

The timeout we spend waiting for the shard witness has been removed. We
don't need this because we wait a whole block time anyway between blocks
and we have a timeout for the propose step individually too. Once shard
0 starts containing data, it's annoying to have this timeout because it
eats into the time available to actually process the data.