feat: Allow attaching orphaned nodes to the merkle trie (#651)
When we write to the trie from parallel threads (like we do in the
replicator), we will write "orphaned" nodes. These are nodes that are
written to the DB, but not yet attached to the root node. This is
because we want to write in parallel threads to different parts of the
trie.

Once the operation is done, we call `attach_to_root` and
`recalculate_hashes` to add new nodes to the trie. This way, we can
parallelize construction of the trie.

Also remove excluded_hash() and snapshot() which we don't use anymore.
The `get_trie_node_metadata` is still supported, which can be used to
get info about the parts of a trie.