feat: Add trie paged iteration + remove attach_to_root (#660)
Add a method that will do paged iteration over the trie. 
- Support paging at a arbitrary sub-tree. 
  - This is used to iterate over all keys in a virtual trie shard("vts")
- Generate page_token/ start from next_page_token so that all keys are
covered
- The page_token is complex Vec<Vec<>>, so encode/decode it to a base64
string
- Tests

Also this PR removes the attach_to_root and the recalculate_hash since
we don't need it anymore - We'll do the iteration on each vts, insert it
into the trie while it is already attached to the root. This way, the
hash is always consistent. So delete all that code