fix: Simplify the replication by removing trie/DB threads (#671)
Simplify the replicator code by removing the trie / DB threads. Instead,
we'll use tokio's Mutex() to lock the DB and Trie (separately) while
processing.

This is effectively the same thing, but is simpler to reason about
because we can just use `.await` everywhere instead of using tokio
threads and tx/rx channels to communicate. Makes the code simpler. This
PR also separates out the "core" logic of commit to DB, commit to Trie
from the parallel threads orchestration to make the code easier to
reason about.