chore: push protocol release out by 2 weeks (#690)
Pushed the protocol release out by 2 weeks. 

We need 1 extra week for testing, pushed by 2 weeks to be conservative
and avoid another upgrade if we end up needing more than a week to test
and make any patches we need. If everything is ready in a week, we can
bump the activation date down and mint a new release.