feat: implement pruning for storage lends (#683)
Prune revokes for storage lends (messages where num_units is set to 0)
as soon as they're merged. We don't want these message to consume
storage and we don't expect conflicts to be frequent so the fact that
these messages are pruned immediately should not be consequential to
users.

In order to protect against reconciling old messages when these revokes
have been pruned, we set the recency window to 1hr for storage lend
messages.