feat: enable migrating onchain events to shard 0 and fix bugs with storage lending (#669)
Add an rpc to migrate onchain events to shard 0 from each of the shard 1
and 2.

In addition, this feature contains several fixes necessary to support
storage lends properly. The bugs were discovered while testing in
testnet. As of this feature storage lending is fully supported in
testnet.

Fixes: 
- Reset event id to the current block number when starting to propose,
validate, or commit a block in the block engine. This was missing so all
event ids were being produced with height 0 and we exhausted the
available sequence numbers.
- Force storage lenders to retain 1 unit of storage so they can revoke
storage they've lent. Otherwise, the lender runs into rate limits when
trying to revoke lent storage after lending all their storage out. This
is easier and safer than bypassing rate limits.
- Listen to confirmed blocks in the mempool and evict storage lends that
have been put into blocks in shard 0.
- Have rate limits in the mempool key off the fid's shard and never
shard 0. The mempool was crashing trying to find a `ShardStore` for
shard 0.
- Add a count of onchain events to the `GetInfo` rpc so we can validate
that the migration of onchain events to shard 0 has happened correctly.