fix: reset events sequence number on validate (#713)
We ran into a case where event id sequence numbers for a height ended up
overflowing because a block with a lot of events (signer revoke) came in
via sync and entered the validate codepath with an existing, high
sequence number for events. Always reset the sequence numbers before
validate to avoid this.