chore: take a replication snapshot if there are none (#734)
The snapshot interval is very infrequent and there are issues when the
bootstrap nodes reboot because it takes them a long time to take their
first snapshot. This feature takes a snapshot if none exist-- this does
mean that the snapshot won't exist on other nodes but it's worth the
tradeoff for now. This snapshot will be the first one pruned anyway.