fix: replication client checks to account for storage lend messages (#735)
Fix the fid/vts check in the replication client to account for storage
lend messages. These messages can appear in a shard that the main fid on
the message doesn't belong to.

Also fixes the code to move the snapshot directory contents into the
main rocksdb directory to work with a dockerized setup.